Core Update 7 (#1899)
* JavaScript and Ajax updates
* Code documentation and bug hunting
* Added a mobile $_SESSION variable to allow changing pages for mobile devices.
* Login screen moves the logo to the bottom on mobile like cockpit does! (makes typing easier on a touch screen)
* Fixed setting button requiring two single clicks to open. (Thanks for this 2016 pull @AshotN)
* Lots of code tweaks to make things more uniform and readable.

## Commits

* Update tools.js
* Update album.class.php
* Update ajax.server.php
* Update ampache.cfg.php.dist
* Update users.php
* method docstrings
* Replce and / or with && / ||
* Update art.class.php
* formatting
* get rid of the 'Closed' text in the editdialog
* missing quotes on some strings
* make generate_thumb only return array
* update browse::show_objects
* codacy updates
* css
* remove some $_GET and $_POST

+++ lib/class/subsonic_api.class.php
@@ -77,7 +77,7 @@ class Subsonic_Api
             $hex    = substr($password, 4);
             $decpwd = '';
             for ($count = 0; $count < strlen($hex); $count += 2) {
-                $decpwd .= chr(hexdec(substr($hex, $count, 2)));
+                $decpwd .= chr((int) hexdec(substr($hex, $count, 2)));
             }
             $password = $decpwd;
         }
@@ -274,7 +274,7 @@ class Subsonic_Api
                     if (!isset($tagsArray[$childTagName])) {
                         //only entry with this key
 
-                        if (count((int) $childProperties) === 0) {
+                        if (count($childProperties) === 0) {
                             $tagsArray[$childTagName] = (object) $childProperties;
                         } elseif (self::has_Nested_Array($childProperties)) {
                             $tagsArray[$childTagName] = (object) $childProperties;
